no segmentation fault, god is good

+++ ft_putnbr_fd.c
@@ -6,21 +6,21 @@
 /*   By: fmoreira <fmoreira@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/27 19:25:47 by fmoreira          #+#    #+#             */
-/*   Updated: 2022/04/27 19:38:23 by fmoreira         ###   ########.fr       */
+/*   Updated: 2022/05/04 20:44:51 by fmoreira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(size_t n, int fd)
 {
-	long int	i;
+	size_t	i;
 
 	i = n;
 	if (n < 0)
 	{
 		write(fd, "-", sizeof(char));
-		i = (long int)n * -1;
+		i = (size_t)n * -1;
 	}
 	n = (i % 10) + '0';
 	i = i / 10;
